deferred loading in pkg/web_app: api client, admin page widgets and package:intl
