+++ pkg/web_app/lib/src/likes.dart
@@ -5,10 +5,10 @@
 import 'dart:async';
 import 'dart:html';
 
-import 'package:intl/intl.dart';
+import 'package:intl/intl.dart' deferred as intl;
 import 'package:mdc_web/mdc_web.dart' show MDCIconButtonToggle;
 
-import 'account.dart';
+import 'api_client/api_client.dart' deferred as api_client;
 import 'page_data.dart';
 
 Future<void> _done = Future.value();
@@ -22,18 +22,19 @@ void setupLikesList() {
   document.querySelectorAll('.-pub-like-button').forEach((likeButton) {
     final package = likeButton.dataset['package']!;
 
-    likeButton.onClick.listen((Event e) {
+    likeButton.onClick.listen((Event e) async {
+      await api_client.loadLibrary();
       final text = likeButton.querySelector('.-pub-like-button-label')!;
       final img =
           likeButton.querySelector('.-pub-like-button-img') as ImageElement;
       if (text.innerText == 'LIKE') {
         text.innerText = 'UNLIKE';
         img.src = likeButton.dataset['thumb_up_filled'];
-        _enqueue(() => client.likePackage(package));
+        _enqueue(() => api_client.client.likePackage(package));
       } else {
         text.innerText = 'LIKE';
         img.src = likeButton.dataset['thumb_up_outlined'];
-        _enqueue(() => client.unlikePackage(package));
+        _enqueue(() => api_client.client.unlikePackage(package));
       }
     });
   });
@@ -52,23 +53,26 @@ void setupLikes() {
   int likesDelta = 0;
 
   // keep in-sync with app/lib/frontend/templates/views/shared/detail/header.dart
-  String likesString() {
+  Future<String> likesString() async {
+    await intl.loadLibrary();
     final likesCount = pageData.pkgData!.likes + likesDelta;
-    return NumberFormat.compact().format(likesCount);
+    return intl.NumberFormat.compact().format(likesCount);
   }
 
-  iconButtonToggle.listen(MDCIconButtonToggle.changeEvent, (Event e) {
+  iconButtonToggle.listen(MDCIconButtonToggle.changeEvent, (Event e) async {
     likeButton.blur();
+    await api_client.loadLibrary();
     if (iconButtonToggle.on ?? false) {
       // The button has shifted to on.
       likesDelta++;
-      likes!.innerText = likesString();
-      _enqueue(() => client.likePackage(pageData.pkgData!.package));
+      likes!.innerText = await likesString();
+      _enqueue(() => api_client.client.likePackage(pageData.pkgData!.package));
     } else {
       // The button has shifted to off.
       likesDelta--;
-      likes!.innerText = likesString();
-      _enqueue(() => client.unlikePackage(pageData.pkgData!.package));
+      likes!.innerText = await likesString();
+      _enqueue(
+          () => api_client.client.unlikePackage(pageData.pkgData!.package));
     }
   });
 }
